WIP: implement callback from a paywall to `RevenueCat` when a font in a paywall fails to load
