+++ RevenueCatUI/Purchasing/PaywallPurchasesType.swift
@@ -11,7 +11,7 @@
 //
 //  Created by Nacho Soto on 9/12/23.
 
-import RevenueCat
+@_spi(Internal) import RevenueCat
 
 /// A simplified protocol for the subset of `PurchasesType` needed for `RevenueCatUI`.
 @available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
@@ -30,6 +30,9 @@ protocol PaywallPurchasesType: Sendable {
 
     @Sendable
     func track(paywallEvent: PaywallEvent) async
+
+    @Sendable
+    func failedToLoadFontWithConfig(_ fontConfig: UIConfig.FontsConfig)
 
 }
