+++ RevenueCatUI/Templates/V2/Components/Text/TextComponentViewModel.swift
@@ -286,33 +286,8 @@ extension TextComponentStyle {
             return GenericFont.sansSerif.makeFont(fontSize: fontSize)
         }
 
-        let customFont = self.resolveFont(size: fontSize, name: name, uiConfigProvider: uiConfigProvider)
+        let customFont = uiConfigProvider.resolveFont(size: fontSize, name: name)
         return customFont ?? GenericFont.sansSerif.makeFont(fontSize: fontSize)
-    }
-
-    static private func resolveFont(
-        size fontSize: CGFloat,
-        name: String,
-        uiConfigProvider: UIConfigProvider
-    ) -> Font? {
-        guard let familyName = uiConfigProvider.getFontFamily(for: name)  else {
-            Logger.warning("Mapping for '\(name)' could not be found. Falling back to system font.")
-            return nil
-        }
-
-        // Check if the family name is a generic font (serif, sans-serif, monospace)
-        if let genericFont = GenericFont(rawValue: familyName) {
-            return genericFont.makeFont(fontSize: fontSize)
-        }
-
-        guard let customFont = UIFont(name: familyName, size: fontSize) else {
-            Logger.warning("Custom font '\(familyName)' could not be loaded. Falling back to system font.")
-            return nil
-        }
-
-        // Apply dynamic type scaling
-        let uiFont = UIFontMetrics.default.scaledFont(for: customFont)
-        return Font(uiFont)
     }
 
 }
